scripts/requirements.txt
pyyaml>=6.0
jsonschema>=4.17
lxml>=4.9
requests>=2.31
